+++ sample/Sample.Orleans.Client.WebApp/ClientApp/app/components/hero/hub-connection.model.ts
@@ -1,6 +1,6 @@
 import { IHubConnectionOptions } from "@aspnet/signalr-client/dist/src/IHubConnectionOptions";
 
-import { Dictionary } from "./core/collection";
+import { Dictionary } from "./utils/dictionary";
 
 export enum ConnectionStatus {
 	connected,
