+++ sample/Sample.Orleans.Client.WebApp/ClientApp/app/components/hero/utils/dictionary.ts
